% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountyGPS.R
\docType{data}
\name{CountyGPS}
\alias{CountyGPS}
\title{CountyGPS}
\format{A data frame with 4 variables:
\describe{
\item{\code{CountyCode}}{County Code}
\item{\code{County}}{County}
\item{\code{Latitude}}{Latitude}
\item{\code{Longitude}}{Longitude}
}}
\usage{
data(CountyGPS)
}
\description{
This data contains the GPS coordinates of each county.
}
\keyword{datasets}
