% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCatalogue.R
\docType{data}
\name{DataCatalogue}
\alias{DataCatalogue}
\title{Dataset Catalogue}
\format{A data frame with 4 variables:
\describe{
\item{\code{Dataset}}{The dataset label e.g V3_T2.3}
\item{\code{Volume}}{The PDF Volume in which the dataset has been published}
\item{\code{Table No in PDFs}}{The table number in the PDF where the data was scrapped }
\item{\code{Dataset Description}}{Description of the data}
}}
\usage{
data(DataCatalogue)
}
\description{
This data shows the description of each of the datasets contained in the package.
}
\keyword{datasets}
