% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KenyaCounties_SHP.R
\docType{data}
\name{KenyaCounties_SHP}
\alias{KenyaCounties_SHP}
\title{KenyaCounties_SHP}
\format{An object of class \code{SpatialPolygonsDataFrame} with 47 rows and 7 columns.}
\usage{
data(KenyaCounties_SHP)
}
\description{
Shapefiles of Kenya County boundaries
}
\details{
Important note: To access this data, ensure you have rgdal and/or sp installed.
}
\keyword{datasets}
