% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.1.R
\docType{data}
\name{V1_T2.1}
\alias{V1_T2.1}
\title{Volume I: Table 2.1}
\format{A data frame with 2 variables:
\describe{
\item{\code{Indicator}}{Indicator}
\item{\code{Value}}{Value}
}}
\usage{
data(V1_T2.1)
}
\description{
Census Indicators at a Glance, 2019
}
\keyword{datasets}
