% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.2.R
\docType{data}
\name{V1_T2.2}
\alias{V1_T2.2}
\title{Volume I: Table 2.2}
\format{A data frame with 5 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Male}}{Number of Males}
\item{\code{Female}}{Number of Females}
\item{\code{Intersex}}{Intersex}
\item{\code{Total}}{Total number of individuals per County}
}}
\usage{
data(V1_T2.2)
}
\description{
Distribution of Population by Sex and County
}
\keyword{datasets}
