% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.3.R
\docType{data}
\name{V1_T2.3}
\alias{V1_T2.3}
\title{Volume I: Table 2.3}
\format{A data frame with 4 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Population}}{Population Size}
\item{\code{NumberOfHouseholds}}{Number of Households}
\item{\code{AverageHouseholdSize}}{Average Household Size}
}}
\usage{
data(V1_T2.3)
}
\description{
Distribution of Population, Number of Households and Average Household Size by County
}
\keyword{datasets}
