% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.4.R
\docType{data}
\name{V1_T2.4}
\alias{V1_T2.4}
\title{Volume I: Table 2.4}
\format{A data frame with 4 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Population}}{Population Size}
\item{\code{LandArea(in Sq. Km)}}{Land Area in Sq. Km}
\item{\code{Population Density(No. per Sq. Km)}}{Population Density (No. per Sq. Km)}
}}
\usage{
data(V1_T2.4)
}
\description{
Distribution of Population, Land Area and Population Density by County
}
\keyword{datasets}
