% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.6.R
\docType{data}
\name{V1_T2.6}
\alias{V1_T2.6}
\title{Volume I: Table 2.6}
\format{A data frame with 6 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether the area is a County or SubCounty}
\item{\code{Population}}{Population Size}
\item{\code{NumberOfHouseholds}}{Number of Households}
\item{\code{AverageHouseholdSize}}{Average Household Size}
}}
\usage{
data(V1_T2.6)
}
\description{
Distribution of Population, Number of Households and Average Household Size by Sub-County
}
\keyword{datasets}
