% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1_T2.7.R
\docType{data}
\name{V1_T2.7}
\alias{V1_T2.7}
\title{Volume I: Table 2.7}
\format{A data frame with 6 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether the area is a County or SubCounty}
\item{\code{Population}}{Population Size}
\item{\code{LandArea(in Sq. Km)}}{Land Area in Sq. Km}
\item{\code{Population Density(No. per Sq. Km)}}{Population Density (No. per Sq. Km)}
}}
\usage{
data(V1_T2.7)
}
\description{
Distribution of Population, Land Area and Population Density by County and SubCounty
}
\keyword{datasets}
