% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_T2.2.R
\docType{data}
\name{V2_T2.2}
\alias{V2_T2.2}
\title{Volume 2: Table 2.2}
\format{A data frame with 10 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Sex_Total}}{Total number of individuals per County}
\item{\code{Sex_Male}}{Number of Males}
\item{\code{Sex_Female}}{Number of Females}
\item{\code{Sex_Intersex}}{Number of Intersex Persons}
\item{\code{Households_Total}}{Total number of households}
\item{\code{Households_Conventional}}{Total number of Conventional households}
\item{\code{Households_GroupQuarters}}{Total number of GroupQuarters households}
\item{\code{LandArea(Sq km)}}{Land Area in Square Kilometres}
\item{\code{Density(Persons per Sq km)}}{Number of Persons per square kilometre**}
}
** Please confirm this definition with KNBS.}
\usage{
data(V2_T2.2)
}
\description{
Distribution of Population by Sex, Number of Households, Land Area, Population Density and County
}
\keyword{datasets}
