% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_T2.5.R
\docType{data}
\name{V2_T2.5}
\alias{V2_T2.5}
\title{Volume 2: Table 2.5}
\format{A data frame with 13 variables:
\describe{
\item{\code{Urban Center}}{Urban Center}
\item{\code{County}}{County}
\item{\code{Total}}{Total number of individuals}
\item{\code{Male}}{Number of Males}
\item{\code{Female}}{Number of Females}
}}
\usage{
data(V2_T2.5)
}
\description{
Distribution of Population by Urban Centres, Sex* and County
}
\keyword{datasets}
