% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T1.1.R
\docType{data}
\name{V3_T1.1}
\alias{V3_T1.1}
\title{Volume 3: Table 1.1}
\format{A data frame with 2 variables:
\describe{
\item{\code{Year}}{Year}
\item{\code{Population (millions)}}{Population in millions}
}}
\usage{
data(V3_T1.1)
}
\description{
Summary of Census Counts in Kenya
}
\keyword{datasets}
