% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T1.2.R
\docType{data}
\name{V3_T1.2}
\alias{V3_T1.2}
\title{Volume 3: Table 1.2}
\format{A data frame with 4 variables:
\describe{
\item{\code{CountyCode}}{County Code (1 - 47)}
\item{\code{County}}{County}
\item{\code{SubCountyCode}}{Sub County Code}
\item{\code{SubCounty}}{Sub County}
}}
\usage{
data(V3_T1.2)
}
\description{
List of Counties and Sub-Counties
}
\keyword{datasets}
