% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T2.1.R
\docType{data}
\name{V3_T2.1}
\alias{V3_T2.1}
\title{Volume 3: Table 2.1}
\format{A data frame with 3 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Sub-Location}}{Sub Location}
\item{\code{Status/Reason}}{Status/Reason}
}}
\usage{
data(V3_T2.1)
}
\description{
Sub-locations with no People on the Census Night by Status/Reason
}
\keyword{datasets}
