% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T2.3.R
\docType{data}
\name{V3_T2.3}
\alias{V3_T2.3}
\title{Volume 3: Table 2.3}
\format{A data frame with 6 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{Age}}{Age}
\item{\code{Male}}{Number of Males}
\item{\code{Female}}{Number of Females}
\item{\code{Total}}{Total number of individuals per County}
}
** Please confirm this definition with KNBS.}
\usage{
data(V3_T2.3)
}
\description{
Distribution of Population by Age, Sex*, County and Sub- County
}
\keyword{datasets}
