% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T2.4a.R
\docType{data}
\name{V3_T2.4a}
\alias{V3_T2.4a}
\title{Volume 3: Table 2.4a}
\format{A data frame with 4 variables:
\describe{
\item{\code{Age}}{Age}
\item{\code{Male}}{Number of Males}
\item{\code{Female}}{Number of Females}
\item{\code{Total}}{Total number of individuals per County}
}}
\usage{
data(V3_T2.4a)
}
\description{
Distribution of Rural Population by Age, Sex* and County
}
\keyword{45}
\keyword{47.}
\keyword{Mombasa.}
\keyword{Nairobi}
\keyword{Please}
\keyword{So}
\keyword{and}
\keyword{counties,}
\keyword{data}
\keyword{datasets}
\keyword{excludes}
\keyword{have}
\keyword{instead}
\keyword{note}
\keyword{of}
\keyword{that}
\keyword{this}
\keyword{we}
