% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V3_T2.4b.R
\docType{data}
\name{V3_T2.4b}
\alias{V3_T2.4b}
\title{Volume 3: Table 2.4b}
\format{A data frame with 5 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Age}}{Age}
\item{\code{Male}}{Number of Males}
\item{\code{Female}}{Number of Females}
\item{\code{Total}}{Total number of individuals per County}
}}
\usage{
data(V3_T2.4b)
}
\description{
Distribution of Urban Population by Age, Sex* and County
}
\keyword{datasets}
