% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T1.1.R
\docType{data}
\name{V4_T1.1}
\alias{V4_T1.1}
\title{Volume 4: Table 1.1}
\format{A data frame with 2 variables:
\describe{
\item{\code{Year}}{Year}
\item{\code{Population (millions)}}{Population in millions}
}}
\usage{
data(V4_T1.1)
}
\description{
Summary of Census Counts in Kenya
}
\keyword{datasets}
