% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T1.9.R
\docType{data}
\name{V4_T1.9}
\alias{V4_T1.9}
\title{Volume 4: Table 1.9}
\format{A data frame with 4 variables:
\describe{
\item{\code{CountyCode}}{County Code (1 - 47)}
\item{\code{County}}{County}
\item{\code{SubCountyCode}}{Sub County Code}
\item{\code{SubCounty}}{Sub County}
}}
\usage{
data(V4_T1.9)
}
\description{
List of Counties and Sub-Counties
}
\keyword{datasets}
