% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.10.R
\docType{data}
\name{V4_T2.10}
\alias{V4_T2.10}
\title{Volume 4: Table 2.10}
\format{A data frame with 12 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{Total}}{Total number of households}
\item{\code{CH}}{Number of Conventional Households}
\item{\code{GQ}}{Number of Group Quarters Households}
\item{\code{CH_Owned_No}}{Number of conventional households, whose main dwelling unit is owned.}
\item{\code{CH_Owned_Perc}}{Percentage of conventional households, whose main dwelling unit is owned.}
\item{\code{CH_Rent/Provided_No}}{Number of conventional households, whose main dwelling unit is rented / provided.}
\item{\code{CH_Rent/Provided_Perc}}{Percentage of conventional households, whose main dwelling unit is rented / provided.}
\item{\code{NotStated_No}}{Number of conventional households, where the tenure status of the main dwelling unit is not stated.}
\item{\code{NotStated_Perc}}{Percentage of conventional households, where the tenure status of the main dwelling unit is not stated.}
}}
\usage{
data(V4_T2.10)
}
\description{
Distribution of Households and Tenure Status of Main Dwelling Unit by Area of Residence,
County and Sub-County
}
\keyword{datasets}
