% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.11a.R
\docType{data}
\name{V4_T2.11a}
\alias{V4_T2.11a}
\title{Volume 4: Table 2.11a}
\format{A data frame with 10 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{Owned}}{Number of conventional households that own their main dwelling unit}
\item{\code{Purchased_No.}}{Number of owned main dwelling units, that are purchased}
\item{\code{Purchased_Perc}}{Percentage of owned main dwelling units, that are purchased}
\item{\code{Constructed_No.}}{Number of owned main dwelling units, that are consructed}
\item{\code{Constructed_Perc}}{Percentage of owned main dwelling units, that are constructed}
\item{\code{Inherited_No.}}{Number of owned main dwelling units, that are inherited}
\item{\code{Inherited_Perc}}{Percentage of owned main dwelling units, that are inherited}
}}
\usage{
data(V4_T2.11a)
}
\description{
Distribution of Households Owning the Main Dwelling Unit by Mode of Acquisition, Area of Residence, County and Sub-County.
}
\keyword{datasets}
