% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.11b.R
\docType{data}
\name{V4_T2.11b}
\alias{V4_T2.11b}
\title{Volume 4: Table 2.11b}
\format{A data frame with 16 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{Rented/Provided}}{Number of conventional households that are Rented/Provided}
\item{\code{...}}{The other variables show the distribution of Rented/Provided dwelling units, by different providers.}
}}
\usage{
data(V4_T2.11b)
}
\description{
Distribution of Households Renting/Provided with the Main Dwelling Unit by Provider, Area of Residence, County and Sub-County.
}
\keyword{datasets}
