% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.12.R
\docType{data}
\name{V4_T2.12}
\alias{V4_T2.12}
\title{Volume 4: Table 2.12}
\format{A data frame with 17 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{ConventionalHouseholds}}{Number of conventional households.}
\item{\code{...}}{The other variables show the distribution of conventional households by dominant roofing material of main dwelling unit.}
}}
\usage{
data(V4_T2.12)
}
\description{
Percentage Distribution of Conventional Households by Dominant Roofing Material of
Main Dwelling Unit, Area of Residence, County and Sub-County
}
\keyword{datasets}
