% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.14.R
\docType{data}
\name{V4_T2.14}
\alias{V4_T2.14}
\title{Volume 4: Table 2.14}
\format{A data frame with 14 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{ConventionalHouseholds}}{Number of conventional households.}
\item{\code{...}}{The other variables show the distribution of conventional households by dominant floor material of main dwelling unit.}
}}
\usage{
data(V4_T2.14)
}
\description{
Percentage Distribution of Conventional Households by Dominant Floor Material of
Main Dwelling Unit, Area of Residence, County and Sub-County
}
\keyword{datasets}
