% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.15.R
\docType{data}
\name{V4_T2.15}
\alias{V4_T2.15}
\title{Volume 4: Table 2.15}
\format{A data frame with 19 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the percentage distribution of conventional households by main source of drinking water.}
}}
\usage{
data(V4_T2.15)
}
\description{
Percentage Distribution of Conventional Households by Main Source of Drinking Water,
Area of Residence, County and Sub-County.
}
\keyword{datasets}
