% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.16.R
\docType{data}
\name{V4_T2.16}
\alias{V4_T2.16}
\title{Volume 4: Table 2.16}
\format{A data frame with 14 variables:

\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the percentage distribution of conventional households by main mode of human waste disposal.}
}}
\usage{
data(V4_T2.16)
}
\description{
Percentage Distribution of Conventional Households by Main Mode of Human Waste Disposal,
Area of Residence, County and Sub-County.
}
\keyword{datasets}
