% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.18.R
\docType{data}
\name{V4_T2.18}
\alias{V4_T2.18}
\title{Volume 4: Table 2.18}
\format{A data frame with 12 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the percentage distribution of conventional households by main type of cooking fuel.}
}}
\usage{
data(V4_T2.18)
}
\description{
Percentage Distribution of Conventional Households by Main Type of Cooking Fuel,
Area of Residence, County and Sub-County.
}
\keyword{datasets}
