% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.19.R
\docType{data}
\name{V4_T2.19}
\alias{V4_T2.19}
\title{Volume 4: Table 2.19}
\format{A data frame with 18 variables

\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the percentage distribution of conventional households by main type of lighting fuel.}
}}
\usage{
data(V4_T2.19)
}
\description{
Percentage Distribution of Conventional Households by Main Type of Lighting Fuel,
Area of Residence, County and Sub-County.
}
\keyword{datasets}
