% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.2.R
\docType{data}
\name{V4_T2.2}
\alias{V4_T2.2}
\title{Volume 4: Table 2.2}
\format{A data frame with 14 variables
\describe{
\item{\code{County}}{County}
\item{\code{Gender}}{Gender}
\item{\code{Total}}{Total}
\item{\code{StillinSchool_No.}}{Number of persons still in school / a learning institution.}
\item{\code{StillinSchool_Perc}}{Percentage of personsstill in school / a learning institution.}
\item{\code{LeftSchoolAfterC_No.}}{Number of persons who left school / a learning institution, after completion.}
\item{\code{LeftSchoolAfterC_Perc}}{Percentage of persons who left school / a learning institution, after completion.}
\item{\code{LeftSchoolBeforeC_No.}}{Number of persons who left school / a learning institution, before completion.}
\item{\code{LeftSchoolBeforeC_Perc}}{Percentage of persons who left school / a learning institution, before completion.}
\item{\code{NeverbeentoSchool_No.}}{Number of persons who have never been to a school / a learning institution.}
\item{\code{NeverbeentoSchool_Perc}}{Percentage of persons who have never been to a school / a learning institution.}
\item{\code{...}}{The other variables indicate situations where the state of school attendance is not known / not stated.}

Intersex population is excluded from the table since it is too small to be distributed at sub-national level.
}}
\usage{
data(V4_T2.2)
}
\description{
Table 2.2: Distribution of Population Age 3 Years and Above by School Attendance Status,
Area of Residence, Sex, County and Sub-County
}
\keyword{datasets}
