% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.20.R
\docType{data}
\name{V4_T2.20}
\alias{V4_T2.20}
\title{Volume 4: Table 2.20}
\format{A data frame with 10 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the number of households carrying out different agricultural practices.}
}}
\usage{
data(V4_T2.20)
}
\description{
Distribution of Households Practicing Agriculture, Fishing and Irrigation by County and Sub County.
}
\keyword{datasets}
