% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.21.R
\docType{data}
\name{V4_T2.21}
\alias{V4_T2.21}
\title{Volume 4: Table 2.21}
\format{A data frame with 11 variables
The other variables show the number of households growing different types of permanent crops.}
\usage{
data(V4_T2.21)
}
\description{
Distribution of Households Growing Permanent Crops by Type and County.
}
\keyword{datasets}
