% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.22.R
\docType{data}
\name{V4_T2.22}
\alias{V4_T2.22}
\title{Volume 4: Table 2.22}
\format{A data frame with 23 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the number of households growing other crops.}
}}
\usage{
data(V4_T2.22)
}
\description{
Distribution of Households Growing Other Crops by Type, County and Sub County.
}
\keyword{datasets}
