% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.23.R
\docType{data}
\name{V4_T2.23}
\alias{V4_T2.23}
\title{Volume 4: Table 2.23}
\format{A data frame with 19 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the number of households with livestock of each type, including beehives, fish ponds and fish cages.}
}}
\usage{
data(V4_T2.23)
}
\description{
Distribution of Households Rearing Livestock and Fish by County and Sub County.
}
\keyword{datasets}
