% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.24.R
\docType{data}
\name{V4_T2.24}
\alias{V4_T2.24}
\title{Volume 4: Table 2.24}
\format{A data frame with 19 variables
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{...}}{The other variables show the number of livestock of each type, including number of beehives, fish ponds and fish cages.}
}}
\usage{
data(V4_T2.24)
}
\description{
Distribution of Livestock Population by Type, Fish Ponds and Fish Cages
by County and Sub County
}
\keyword{datasets}
