% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.25.R
\docType{data}
\name{V4_T2.25}
\alias{V4_T2.25}
\title{Volume 4: Table 2.25}
\format{A data frame with 9 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or Sub County}
\item{\code{LandSize}}{Area of Agricultural land (Hectares)}
\item{\code{LandSize_Subsistence}}{Area of agricutural land by main purpose (Hectares):Subsistence}
\item{\code{LandSize_Commercial}}{Area of agricutural land by main purpose (Hectares):Commercial}
\item{\code{No.FHS}}{Total number of Farming Households}
\item{\code{No.FHS_Subsistence}}{Number of farming households by main purpose of agricultural production:Subsistence}
\item{\code{No.FHS_Commercial}}{Number of farming households by main purpose of agricultural production:Commercial}
}}
\usage{
data(V4_T2.25)
}
\description{
Distribution of area (hectares) of Agricultural land and Farming Households by
purpose of production, County and Sub-County
}
\keyword{datasets}
