% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.26.R
\docType{data}
\name{V4_T2.26}
\alias{V4_T2.26}
\title{Volume 4: Table 2.26}
\format{A data frame with 16 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or SubCounty}
\item{\code{PopAge>5_Total}}{Total number of persons Aged 5 years and above}
\item{\code{PopAge>5_Male}}{Total number of males Aged 5 years and above}
\item{\code{PopAge>5_Female}}{Total number of females Aged 5 years and above}
\item{\code{WithDisability_Total}}{Total number of persons Aged 5 years and above, With Disability}
\item{\code{WithDisability_Male}}{Total number of males Aged 5 years and above, With Disability}
\item{\code{WithDisability_Female}}{Total number of females Aged 5 years and above, With Disability}
\item{\code{WithoutDisability_Total}}{Total number of persons Aged 5 years and above, Without Disability}
\item{\code{WithoutDisability_Male}}{Total number of males Aged 5 years and above, Without Disability}
\item{\code{WithoutDisability_Female}}{Total number of females Aged 5 years and above, Without Disability}
\item{\code{NotStated_Total}}{Total number of persons Aged 5 years and above, whose Disability is Not Stated}
\item{\code{NotStated_Male}}{Total number of males Aged 5 years and above,  whose Disability is Not Stated}
\item{\code{NotStated_Female}}{Total number of females Aged 5 years and above,  whose Disability is Not Stated}
\item{\code{PercentageWithDisability}}{Percentage number of persons Aged 5 years and above, With Disability}
}}
\usage{
data(V4_T2.26)
}
\description{
Distribution of Population aged 5 years and above by Disability Status, Sex, Area of Residence, County and Sub-County
}
\keyword{57}
\keyword{For}
\keyword{Intersex}
\keyword{at}
\keyword{be}
\keyword{datasets}
\keyword{disability}
\keyword{distributed}
\keyword{excluded}
\keyword{from}
\keyword{had}
\keyword{intersex}
\keyword{is}
\keyword{it}
\keyword{level,}
\keyword{level.}
\keyword{national}
\keyword{population}
\keyword{since}
\keyword{small}
\keyword{sub-national}
\keyword{table}
\keyword{the}
\keyword{to}
\keyword{too}
