% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.27.R
\docType{data}
\name{V4_T2.27}
\alias{V4_T2.27}
\title{Volume 4: Table 2.27}
\format{A data frame with 21 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an area is a County or SubCounty}
\item{\code{Vision_Total}}{Total number of persons with Vision disability}
\item{\code{Vision_Male}}{Total number of males with Vision disability}
\item{\code{Vision_Female}}{Total number of females with Vision disability}
\item{\code{Hearing_Total}}{Total number of persons with Hearing disability}
\item{\code{Hearing_Male}}{Total number of males with Hearing disability}
\item{\code{Hearing_Female}}{Total number of females with Hearing disability}
\item{\code{Mobility_Total}}{Total number of persons with Mobility disability}
\item{\code{Mobility_Male}}{Total number of males with Mobility disability}
\item{\code{Mobility_Female}}{Total number of females with Mobility disability}
\item{\code{Cognition_Total}}{Total number of persons with Cognition disability}
\item{\code{Cognition_Male}}{Total number of males with Cognition disability}
\item{\code{Cognition_Female}}{Total number of females with Cognition disability}
\item{\code{SelfCare_Total}}{Total number of persons with SelfCare disability}
\item{\code{SelfCare_Male}}{Total number of males with SelfCare disability}
\item{\code{SelfCare_Female}}{Total number of females with SelfCare disability}
\item{\code{Communication_Total}}{Total number of persons with Communication disability}
\item{\code{Communication_Male}}{Total number of males with Communication disability}
\item{\code{Communication_Female}}{Total number of females with Communication disability}
}}
\usage{
data(V4_T2.27)
}
\description{
Distribution of Persons with Disability by Type of Disability, Sex 1 , Area of Residence, County and Sub County
}
\keyword{datasets}
