% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.28.R
\docType{data}
\name{V4_T2.28}
\alias{V4_T2.28}
\title{Volume 4: Table 2.28}
\format{A data frame with 9 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{Shows whether an Area is a County or SubCounty}
\item{\code{TP_Total}}{Total number of persons}
\item{\code{TP_Male}}{Total number of males}
\item{\code{TP_Female}}{Total number of females}
\item{\code{PWA_Total}}{Total number of persons with albinism}
\item{\code{PWA_Male}}{Total number of males with albinism}
\item{\code{PWA_Female}}{Total number of females with albinism}
}}
\usage{
data(V4_T2.28)
}
\description{
Distribution of Persons with Albinism by Sex 1 , Area of Residence, County and Sub County
}
\keyword{datasets}
