% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.29.R
\docType{data}
\name{V4_T2.29}
\alias{V4_T2.29}
\title{Volume 4: Table 2.29}
\format{A data frame with 4 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Total}}{Total number of persons}
\item{\code{Male}}{Total number of males}
\item{\code{Female}}{Total number of females}
}}
\usage{
data(V4_T2.29)
}
\description{
Population of Street Persons/Outdoor Sleepers by Sex, Area of Residence and County
}
\keyword{Intersex}
\keyword{Note:}
\keyword{The}
\keyword{and}
\keyword{at}
\keyword{be}
\keyword{datasets}
\keyword{distributed}
\keyword{do}
\keyword{education}
\keyword{enumerated}
\keyword{excluded}
\keyword{figures}
\keyword{from}
\keyword{in}
\keyword{include}
\keyword{institutions}
\keyword{is}
\keyword{it}
\keyword{level.}
\keyword{not}
\keyword{population}
\keyword{programmes}
\keyword{rehabiliation}
\keyword{since}
\keyword{small}
\keyword{sub-national}
\keyword{table}
\keyword{the}
\keyword{those}
\keyword{to}
\keyword{too}
\keyword{undertaking}
\keyword{while}
