% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.3.R
\docType{data}
\name{V4_T2.3}
\alias{V4_T2.3}
\title{Volume 4: Table 2.3}
\format{A data frame with 12 variables, showing the disribution of the current level
of education, by Gender and Sub County.

Intersex population is excluded from the table since it is too small to be distributed at sub-national level.}
\usage{
data(V4_T2.3)
}
\description{
Distribution of Population Age 3 Years and Above Currently Attending School/Learning Institution by Education Level,
Area of Residence, Sex, County and Sub-County
}
\keyword{datasets}
