% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.30.R
\docType{data}
\name{V4_T2.30}
\alias{V4_T2.30}
\title{Volume 4: Table 2.30}
\format{A data frame with 15 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Total}}{Total number of persons per county}
\item{\code{...}}{The other variables show the distribution of different religions in Kenya, including "Don't Know" and "NotStated" options.}
}}
\usage{
data(V4_T2.30)
}
\description{
Distribution of Population by Religious Affiliation and County
}
\keyword{datasets}
