% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.33.R
\docType{data}
\name{V4_T2.33}
\alias{V4_T2.33}
\title{Volume 4: Table 2.33}
\format{A data frame with 18 variables:
\describe{
\item{\code{County}}{County}
\item{\code{SubCounty}}{Sub County}
\item{\code{AdminArea}}{County or SubCounty}
\item{\code{Total}}{Total number of persons Age 3 years and Above}
\item{\code{Male}}{Total number of males Age 3 years and Above}
\item{\code{Female}}{Total number of females Age 3 years and Above}
\item{\code{UoI_Total}}{Total number of persons Age 3 years and Above who Used Internet}
\item{\code{UoI_Total_Perc}}{Percentage of persons Age 3 years and Above who Used Internet}
\item{\code{UoI_Male}}{Total number of males Age 3 years and Above who Used Internet}
\item{\code{UoI_Male_Perc}}{Percentage of males Age 3 years and Above who Used Internet}
\item{\code{UoI_Female}}{Total number of females Age 3 years and Above who Used Internet}
\item{\code{UoI_Female_Perc}}{Percentage of females Age 3 years and Above who Used Internet}
\item{\code{UoDLT_Total}}{Total number of persons Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
\item{\code{UoDLT_Total_Perc}}{Percentage of persons Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
\item{\code{UoDLT_Male}}{Total number of males Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
\item{\code{UoDLT_Male_Perc}}{Percentage of males Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
\item{\code{UoDLT_Female}}{Total number of females Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
\item{\code{UoDLT_Female_Perc}}{Percentage of females Age 3 years and Above who Used Desktop Computer/Laptop/Tablet}
}}
\usage{
data(V4_T2.33)
}
\description{
Distribution of Population Age 3 Years and Above Using Internet and Computer/Laptop/Tablet by
Area of Residence, Sex, County and Sub-County
}
\keyword{(Rural-}
\keyword{(Rural-102,}
\keyword{(Rural-44,}
\keyword{1,298}
\keyword{177}
\keyword{240);}
\keyword{3}
\keyword{342}
\keyword{545);}
\keyword{753,}
\keyword{Intersex}
\keyword{Urban-}
\keyword{Urban-133)}
\keyword{above}
\keyword{aged}
\keyword{and}
\keyword{at}
\keyword{be}
\keyword{computer}
\keyword{datasets}
\keyword{distributed}
\keyword{excluded}
\keyword{from}
\keyword{internet}
\keyword{is}
\keyword{it}
\keyword{level.}
\keyword{of}
\keyword{population}
\keyword{since}
\keyword{small}
\keyword{sub-national}
\keyword{table}
\keyword{the}
\keyword{to}
\keyword{too}
\keyword{use}
\keyword{were}
\keyword{years}
