% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.38.R
\docType{data}
\name{V4_T2.38}
\alias{V4_T2.38}
\title{Volume 4: Table 2.38}
\format{A data frame with 7 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Total}}{Total}
\item{\code{HealthFacility}}{Number of births that occured in a health facility}
\item{\code{Non_HealthFacility}}{Number of births that occured in a non health facility}
\item{\code{Don’t_Know}}{Don’t Know}
\item{\code{Not_Stated}}{Not Stated}
\item{\code{Percent_in_HealthFacility}}{Percentage of births that occured in a health facility}
}}
\usage{
data(V4_T2.38)
}
\description{
Births in the Last 5 Years* by place of Occurrence and County
}
\details{
These are births that occurred in the 5 years preceding the Census
}
\keyword{datasets}
