% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.39.R
\docType{data}
\name{V4_T2.39}
\alias{V4_T2.39}
\title{Volume 4: Table 2.39}
\format{A data frame with 7 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Total}}{Total}
\item{\code{Notified}}{Number of births that were notified}
\item{\code{Not_Notified}}{Number of births that were not Notified}
\item{\code{Don’t_Know}}{Don’t Know}
\item{\code{Not_Stated}}{Not Stated}
\item{\code{Percent_Notified}}{Percentage of births that were notified}
}}
\usage{
data(V4_T2.39)
}
\description{
Notified Births in the Last 12 months by County
}
\details{
These are births that occurred in the 12 months preceding the Census
}
\keyword{datasets}
