% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.4.R
\docType{data}
\name{V4_T2.4}
\alias{V4_T2.4}
\title{Volume 4: Table 2.4}
\format{A data frame with 12 variables, showing the disribution of the highest level
of education reached, by Gender and Sub County.

Intersex population is excluded from the table since it is too small to be distributed at sub-national level.}
\usage{
data(V4_T2.4)
}
\description{
Distribution of Population Age 3 Years and Above by Highest Level of Education Reached, Area of Residence, Sex, County and Sub-County
}
\keyword{datasets}
