% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.40.R
\docType{data}
\name{V4_T2.40}
\alias{V4_T2.40}
\title{Volume 4: Table 2.40}
\format{A data frame with 7 variables:
\describe{
\item{\code{County}}{County}
\item{\code{Total}}{Total}
\item{\code{Notified}}{Number of births that were notified}
\item{\code{Not_Notified}}{Number of births that were not notified}
\item{\code{Don’t_Know}}{Don’t Know}
\item{\code{Not_Stated}}{Not Stated}
\item{\code{Percent_Notified}}{Percentage of births that were notified}
}}
\usage{
data(V4_T2.40)
}
\description{
Notified Births in the Last 5 Years by County
}
\keyword{datasets}
