% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.5.R
\docType{data}
\name{V4_T2.5}
\alias{V4_T2.5}
\title{Volume 4: Table 2.5}
\format{A data frame with 13 variables, showing the disribution of the highest level
of education compleTed, by Gender and Sub County.

Intersex population is excluded from the table since it is too small to be distributed at sub-national level.}
\usage{
data(V4_T2.5)
}
\description{
Distribution of Population Age 3 Years and Above by Highest Level of Education Completed, Area of Residence, Sex, County and Sub-County
}
\keyword{datasets}
