% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.6a.R
\docType{data}
\name{V4_T2.6a}
\alias{V4_T2.6a}
\title{Volume 4: Table 2.6a}
\format{A data frame with 10 variables}
\usage{
data(V4_T2.6a)
}
\description{
Distribution of Population Age 3 Years and Above by School Attendance Status,
Sex and Special Age Groups
}
\keyword{datasets}
