% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.7.R
\docType{data}
\name{V4_T2.7}
\alias{V4_T2.7}
\title{Volume 4: Table 2.7}
\format{A data frame with 5 variables
\describe{
\item{\code{AreaofTraining}}{Area of Training}
\item{\code{Total}}{Total number of persons}
\item{\code{Male}}{Number of males}
\item{\code{Female}}{Number of females}
\item{\code{Intersex}}{Number of intersex persons}
}}
\usage{
data(V4_T2.7)
}
\description{
Distribution of Population Age 15 years and Above by Sex and Main Training Acquired and Qualified for
}
\keyword{datasets}
