% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V4_T2.8b.R
\docType{data}
\name{V4_T2.8b}
\alias{V4_T2.8b}
\title{Volume 4: Table 2.8b}
\format{A data frame with 7 variables
\describe{
\item{\code{SubCounty}}{Sub County}
\item{\code{Gender}}{Gender}
\item{\code{Total}}{Total}
\item{\code{PILF_Working}}{People In Labor Force that are Working}
\item{\code{PILF_SeekingWork/NoWorkAvailable}}{People In Labor Force that are seeking work/no work is available}
\item{\code{POLF}}{People Outside the Labor Force}
\item{\code{NotStated}}{Activity Status Not Stated}
Intersex population is excluded from the table since it is too small to be distributed at sub-national level.

Persons outside the labour force is what was formally referred to as the economically inactive and includes: full-time students,home makers, the retired,
incapacitated persons and those who are either too young or too old to work
}}
\usage{
data(V4_T2.8b)
}
\description{
Distribution of Urban Population Age 5 Years and above by Activity Status, Sex,
County and Sub-County: Persons in the Labour Force
}
\keyword{datasets}
