% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_T1.2.R
\docType{data}
\name{V2_T1.2}
\alias{V2_T1.2}
\title{Volume 2: Table 1.2}
\format{A data frame with 4 variables:
\describe{
\item{\code{CountyCode}}{County Code (1 - 47)}
\item{\code{County}}{County}
\item{\code{SubCountyCode}}{Sub County Code}
\item{\code{SubCounty}}{Sub County}
}}
\usage{
data(V2_T1.2)
}
\description{
A list of Counties and their respective Sub Counties
}
\keyword{datasets}
