% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_T2.3.R
\docType{data}
\name{V2_T2.3}
\alias{V2_T2.3}
\title{Volume 2: Table 2.3}
\format{A data frame with 9 variables:
\describe{
\item{\code{County/SubCounty}}{Specifies the County or SubCounty}
\item{\code{Sex_Total}}{Total number of individuals}
\item{\code{Sex_Male}}{Number of Males}
\item{\code{Sex_Female}}{Number of Females}
\item{\code{Households_Total}}{Total number of households}
\item{\code{Households_Conventional}}{Total number of Conventional households}
\item{\code{Households_GroupQuarters}}{Total number of GroupQuarters households}
\item{\code{LandArea(Sq km)}}{Land Area in Square Kilometres}
\item{\code{Density(Persons per Sq km)}}{Number of Persons per square kilometre**}
}
** Please confirm this definition with KNBS.}
\usage{
data(V2_T2.3)
}
\description{
Distribution of Population by Sex, Number of Households,Land Area, Population Density and Sub County
}
\keyword{datasets}
